+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Fix bug where home timeline toots were being sorted backwards before being truncated
 
 ### v0.48.8
 * Refactor `Toot.contentWithCard()` method